+++ rascal-vscode-extension/src/lsp/RascalLSPConnection.ts
@@ -194,6 +194,14 @@ function gb(amount: integer) {
 }
 
 function calculateRascalMemoryReservation() {
+    const config = vscode.workspace.getConfiguration('rascal.lSP');
+    if (config.has('maxHeapSize')) {
+        const maxHeapSize = config.get('maxHeapSize');
+        if (maxHeapSize !== null) {
+            return `-Xmx${maxHeapSize}M`;
+        }
+    }
+
     // rascal lsp needs at least 800M but runs better with 2G or even 2.5G (especially the type checker)
     if (os.totalmem() >= gb(32)) {
         return "-Xmx2500M";
@@ -209,6 +217,14 @@ function calculateRascalMemoryReservation() {
 }
 
 function calculateDSLMemoryReservation(_dedicated: boolean) {
+    const config = vscode.workspace.getConfiguration('rascal.parametric.lSP');
+    if (config.has('maxHeapSize')) {
+        const maxHeapSize = config.get('maxHeapSize');
+        if (maxHeapSize !== null) {
+            return `-Xmx${maxHeapSize}M`;
+        }
+    }
+
     // this is a hard one, if you register many DSLs, it can grow quite a bit
     // 400MB per language is a reasonable estimate (for average sized languages)
     if (os.totalmem() >= gb(32)) {